refactor: move MDX to /blog
