+++ src/pages/blog/experimenting-react-suspense-with-swr.mdx
