+++ src/pages/blog/index.tsx
@@ -1,224 +1,87 @@
-import { getBlogLink, getDateStr, postIsVisible } from '../../lib/blog-helpers';
-import getBlogIndex, { Blog } from '../../lib/notion/getBlogIndex';
+import * as React from 'react';
 import Image from 'next/image';
-import { GetStaticProps } from 'next';
-import { PostPreview } from '../../lib/notion/getPostPreview';
-import { fetchNotionAsset } from '../../lib/apis/notion/assetAPI';
-import { BlogTag } from '../../components/pages/blog/BlogTag';
+import { blogMeta, BlogMeta } from '../../blogs/meta';
 import { Card } from '../../components/utils/Card';
-import { PreviewNote } from '../../components/pages/blog/PreviewNote';
 import { Page } from '../../components/utils/Page';
+import { getDateString } from '../../lib/utils/date';
+import { BlogTag } from '../../components/pages/blog/BlogTag';
+import { GetStaticProps } from 'next';
+import { sortByDate } from '../../lib/utils/sorting';
 
-type PreviewContent = PostPreview[0] & { source: string | null };
+const PostCard: React.FC<BlogMeta> = props => {
+  const { title, slug, date, tags, published, image } = props;
 
-type Post = Omit<Blog, 'Tags' | 'preview'> & {
-  preview?: PreviewContent[];
-  Tags: string[];
+  return (
+    <Card
+      tag={tags[0]}
+      href="/blog/[slug]"
+      aria-label={title}
+      as={`/blog/${slug}`}
+    >
+      <div>
+        {image && (
+          <Image
+            src={image}
+            width={480}
+            height={320}
+            alt=""
+            role="presentation"
+            className="placeholder object-cover"
+          />
+        )}
+        <div className="mt-2 mb-4 mx-4">
+          <h2 className="text-xl font-bold">{title}</h2>
+          {date && (
+            <div className="mt-1 text-base text-gray-500 dark:text-gray-300">
+              {getDateString(date)}
+            </div>
+          )}
+          {tags.length > 0 && (
+            <div className="blog-tag-group mt-2">
+              {!published && <BlogTag text="draft" />}
+              {tags.map(tag => (
+                <BlogTag text={tag} key={tag} />
+              ))}
+            </div>
+          )}
+        </div>
+      </div>
+    </Card>
+  );
 };
 
-type PostIndexProps = {
-  posts: Post[];
-  preview: boolean;
+type BlogIndexPageProps = {
+  blogsMeta: BlogMeta[];
 };
 
-export const getStaticProps: GetStaticProps<PostIndexProps> = async params => {
-  const { preview } = params;
-  const postsTable = await getBlogIndex();
-
-  const posts: Post[] = Object.keys(postsTable)
-    .map(slug => {
-      const post = postsTable[slug];
-      // remove draft posts in production
-      if (!preview && !postIsVisible(post)) {
-        return null;
-      }
-      return {
-        ...post,
-        Tags: post.Tags.split(','),
-        preview: post.preview.map(pre => ({
-          ...pre,
-          source: null,
-        })),
-      };
-    })
-    .filter(Boolean)
-    .sort((postA, postB) => Math.sign(postB.Date - postA.Date));
-
-  await Promise.all(
-    posts.map(post =>
-      Promise.all(
-        post.preview?.map(async preview => {
-          if (preview.type === 'image') {
-            const source = await fetchNotionAsset(
-              preview.content[0][0] as string,
-              preview.id,
-            );
-
-            preview.source = source;
-          }
-        }),
-      ),
-    ),
+export const getStaticProps: GetStaticProps<BlogIndexPageProps> = async context => {
+  const meta = sortByDate(
+    context.preview
+      ? Object.values(blogMeta)
+      : Object.values(blogMeta).filter(meta => meta.published),
+    'date',
   );
 
   return {
     props: {
-      preview: preview || false,
-      posts,
+      blogsMeta: meta,
     },
-    revalidate: 10,
   };
 };
 
-const FeaturedPostCard: React.FC<{
-  post: Post;
-}> = props => {
-  const { post } = props;
+const BlogIndexPage: React.FC<BlogIndexPageProps> = props => {
+  const { blogsMeta } = props;
 
   return (
-    <Card
-      tag={post.Tags[0]}
-      href="/blog/[slug]"
-      aria-label={post.Page}
-      as={getBlogLink(post.Slug)}
-    >
-      <div className="flex flex-col sm:flex-row flex-wrap sm:items-center">
-        {(post.preview || [])
-          .filter(block => block.type === 'image')
-          .map((block, idx) => {
-            if (idx > 0) {
-              return null;
-            }
-
-            if (!block.content || !block.source) {
-              return null;
-            }
-
-            return (
-              <div className="flex-1 featured-card-flex" key={block.id}>
-                <Image
-                  src={block.source}
-                  width={480}
-                  height={360}
-                  alt=""
-                  role="presentation"
-                  key={block.id}
-                  className="placeholder object-cover"
-                />
-              </div>
-            );
-          })}
-        <div className="flex-1 m-4">
-          <h2 className="text-xl sm:text-2xl font-bold">{post.Page}</h2>
-          {post.Date && (
-            <div className="mt-1 text-base sm:text-lg text-gray-500 dark:text-gray-300">
-              {getDateStr(post.Date)}
-            </div>
-          )}
-          {post.Tags.length > 0 && (
-            <div className="blog-tag-group mt-2">
-              {post.Published !== 'Yes' && <BlogTag text="Draft" />}
-              {post.Tags.map(tag => (
-                <BlogTag text={tag} key={tag} />
-              ))}
-            </div>
-          )}
-        </div>
-      </div>
-    </Card>
-  );
-};
-
-const NormalPostCard: React.FC<{
-  post: Post;
-}> = props => {
-  const { post } = props;
-
-  return (
-    <Card
-      tag={post.Tags[0]}
-      href="/blog/[slug]"
-      aria-label={post.Page}
-      as={getBlogLink(post.Slug)}
-    >
-      <div>
-        {(post.preview || [])
-          .filter(block => block.type === 'image')
-          .map((block, idx) => {
-            if (idx > 0) {
-              return null;
-            }
-
-            if (!block.content || !block.source) {
-              return null;
-            }
-
-            return (
-              <Image
-                src={block.source}
-                width={480}
-                height={360}
-                alt=""
-                role="presentation"
-                key={block.id}
-                className="placeholder object-cover"
-              />
-            );
-          })}
-        <div className="mt-2 mb-4 mx-4">
-          <h2 className="text-xl font-bold">{post.Page}</h2>
-          {post.Date && (
-            <div className="mt-1 text-base text-gray-500 dark:text-gray-300">
-              {getDateStr(post.Date)}
-            </div>
-          )}
-          {post.Tags.length > 0 && (
-            <div className="blog-tag-group mt-2">
-              {post.Published !== 'Yes' && <BlogTag text="Draft" />}
-              {post.Tags.map(tag => (
-                <BlogTag text={tag} key={tag} />
-              ))}
-            </div>
-          )}
-          <p>
-            {(!post.preview || post.preview.length === 0) &&
-              'No preview available'}
-          </p>
-        </div>
-      </div>
-    </Card>
-  );
-};
-
-const PostIndex: React.FC<PostIndexProps> = props => {
-  const { posts, preview } = props;
-
-  const featuredPostIndex = posts.findIndex(post => post.Featured === 'Yes');
-  const featuredPost = featuredPostIndex > -1 ? posts[featuredPostIndex] : null;
-  const normalPosts = posts.filter((_, index) => index !== featuredPostIndex);
-
-  return (
-    <Page titlePre="Blog" className="max-w-screen-lg mx-auto px-6">
-      {preview && (
-        <PreviewNote clearLink="/api/clear-preview" className="mb-8" />
-      )}
+    <Page titlePre="Blog" className="max-w-6xl mx-auto px-6">
       <h1 className="visually-hidden">Blog</h1>
-      {posts.length === 0 && (
-        <p className="text-center">There are no posts yet</p>
-      )}
-      {featuredPost && (
-        // Hack: for showing focus-visible outline
-        <div className="grid grid-cols-1">
-          <FeaturedPostCard post={featuredPost} />
-        </div>
-      )}
-      <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-4 mt-4 content-center">
-        {normalPosts.map(post => (
-          <NormalPostCard post={post} key={post.Slug} />
+      <div className="grid grid-cols-1 xs:grid-cols-2 lg:grid-cols-3 gap-4 mt-4 content-center items-stretch">
+        {blogsMeta.map(meta => (
+          <PostCard {...meta} key={meta.slug} />
         ))}
       </div>
     </Page>
   );
 };
 
-export default PostIndex;
+export default BlogIndexPage;
